Use sidekiq retries
