+++ app/jobs/event_jobs.rb
@@ -19,14 +19,10 @@ module EventJobs
   class Emit < ApplicationJob
     def perform(event_id)
       event = Event.find(event_id)
-      begin
-        Forms::Events::Emit.new(
-          event: event
-        ).save.raise_if_error
-      rescue StandardError => e
-        Forms::Events::ScheduleRetry.new(event: event).save.raise_if_error
-        raise e
-      end
+
+      Forms::Events::Emit.new(
+        event: event
+      ).save.raise_if_error
     end
   end
 end
