--- db/migrate/20190926185450_add_retries_to_events.rb
@@ -1,5 +1,0 @@
-class AddRetriesToEvents < ActiveRecord::Migration[5.2]
-  def change
-    add_column :events, :retries, :integer
-  end
-end
