+++ db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 2019_09_26_185450) do
+ActiveRecord::Schema.define(version: 2019_09_19_162249) do
 
   # These are extensions that must be enabled in order to support this database
   enable_extension "plpgsql"
@@ -31,7 +31,6 @@ ActiveRecord::Schema.define(version: 2019_09_26_185450) do
     t.string "organization_id"
     t.datetime "created_at", null: false
     t.datetime "updated_at", null: false
-    t.integer "retries"
     t.index ["event_object_id", "event_object_type"], name: "index_events_on_event_object_id_and_event_object_type"
     t.index ["organization_id"], name: "index_events_on_organization_id"
     t.index ["type"], name: "index_events_on_type"
